requirements.txt
alabaster==0.7.12
ansible==2.9.3
appdirs==1.4.3
asn1crypto==0.24.0
awxkit==9.2.0
Babel==2.5.1
bcrypt==3.1.6
cffi==1.11.5
chardet==3.0.4
cloud-init==18.5
configobj==5.0.6
cryptography==2.3
dbus-python==1.2.4
decorator==4.2.1
distlib==0.3.0
distro==1.4.0
dnspython==1.16.0
docutils==0.16
ethtool==0.14
file-magic==0.3.0
filelock==3.0.12
gpg==1.10.0
idna==2.5
imagesize==1.2.0
importlib-metadata==1.5.0
importlib-resources==1.1.0
iniparse==0.4
insights-client==3.0.8
Jinja2==2.10.1
jmespath==0.9.0
jsonpatch==1.21
jsonpointer==1.10
jsonschema==2.6.0
M2Crypto==0.35.2
MarkupSafe==0.23
meld3==2.0.0
msgpack==0.6.2
mysql-connector-python==8.0.19
netifaces==0.10.6
oauthlib==2.1.0
packaging==20.1
paramiko==2.4.3
pciutils==2.3.6
perf==0.1
ply==3.9
prettytable==0.7.2
protobuf==3.6.1
psutil==5.6.3
psycopg2==2.7.5
pyasn1==0.3.7
pycparser==2.14
pycurl==7.43.0.2
Pygments==2.5.2
pygobject==3.28.3
pyinotify==0.9.6
PyJWT==1.6.1
PyMySQL==0.8.0
PyNaCl==1.3.0
pyOpenSSL==18.0.0
pyparsing==2.4.6
pyserial==3.1.1
PySocks==1.6.8
python-dateutil==2.6.1
python-dmidecode==3.12.2
python-linux-procfs==0.6
pytz==2017.2
pyudev==0.21.0
PyYAML==3.12
pyzmq==17.0.0
requests==2.20.0
rhnlib==2.8.6
rpm==4.14.2
salt==3000
schedutils==0.6
selinux==2.9
sepolicy==1.1
setools==4.2.2
six==1.11.0
snowballstemmer==2.0.0
sos==3.7
Sphinx==2.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-autoprogram==0.1.5
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
subscription-manager==1.25.17
supervisor==4.0.4
syspurpose==1.25.17
systemd-python==234
urllib3==1.24.2
virtualenv==20.0.7
zipp==3.0.0
